LoggerStore cleanup
