+++ Sources/Pulse/Helpers/ImageProcessor.swift
@@ -1,0 +1,81 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2020–2023 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+#if !os(macOS)
+import UIKit.UIImage
+/// Alias for `UIImage`.
+typealias PlatformImage = UIImage
+#else
+import AppKit.NSImage
+/// Alias for `NSImage`.
+typealias PlatformImage = NSImage
+#endif
+
+#if os(watchOS)
+import ImageIO
+#endif
+
+enum Graphics {
+    /// Creates an image thumbnail. Uses significantly less memory than other options.
+    static func makeThumbnail(from data: Data, targetSize: CGFloat) -> PlatformImage? {
+        guard let source = CGImageSourceCreateWithData(data as CFData, [kCGImageSourceShouldCache: false] as CFDictionary) else {
+            return nil
+        }
+        let options = [
+            kCGImageSourceCreateThumbnailFromImageAlways: true,
+            kCGImageSourceCreateThumbnailFromImageIfAbsent: true,
+            kCGImageSourceCreateThumbnailWithTransform: true,
+            kCGImageSourceThumbnailMaxPixelSize: targetSize] as [CFString: Any]
+        guard let image = CGImageSourceCreateThumbnailAtIndex(source, 0, options as CFDictionary) else {
+            return nil
+        }
+        return PlatformImage(cgImage: image)
+    }
+
+    static func encode(_ image: PlatformImage) -> Data? {
+        guard let source = image.cgImage else {
+            return nil
+        }
+        let data = NSMutableData()
+#if os(watchOS)
+        let type: String = "public.jpeg"
+#else
+        let type: String = "public.heic"
+#endif
+        guard let destination = CGImageDestinationCreateWithData(data as CFMutableData, type as CFString, 1, nil) else {
+            return nil
+        }
+        let options: NSDictionary = [
+            kCGImageDestinationLossyCompressionQuality: 0.33
+        ]
+        CGImageDestinationAddImage(destination, source, options)
+        CGImageDestinationFinalize(destination)
+        guard !data.isEmpty else { return nil }
+        return data as Data
+    }
+
+    static func makeMetadata(from data: Data) -> [String: String]? {
+        guard let image = PlatformImage(data: data) else {
+            return nil
+        }
+        return [
+            "ResponsePixelWidth": String(Int(image.size.width)),
+            "ResponsePixelHeight": String(Int(image.size.height))
+        ]
+    }
+}
+
+#if os(macOS)
+extension NSImage {
+    var cgImage: CGImage? {
+        cgImage(forProposedRect: nil, context: nil, hints: nil)
+    }
+
+    convenience init(cgImage: CGImage) {
+        self.init(cgImage: cgImage, size: .zero)
+    }
+}
+#endif
